# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Learnup",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/teacher/course": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teacher"
                ],
                "summary": "Create a new course by teacher",
                "parameters": [
                    {
                        "description": " ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ladmakhi81_learnup_internals_teacher_dto_req.CreateCourseReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ladmakhi81_learnup_internals_teacher_dto_res.CreateCourseRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ApiError"
                        }
                    }
                }
            }
        },
        "/teacher/courses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teacher"
                ],
                "summary": "Get teacher's courses",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/types.PaginationRes"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "row": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/github_com_ladmakhi81_learnup_internals_teacher_dto_res.FetchCourseItemRes"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ladmakhi81_learnup_internals_teacher_dto_req.CreateCourseReq": {
            "type": "object",
            "required": [
                "abilityToAddComment",
                "canHaveDiscount",
                "categoryId",
                "courseLevel",
                "description",
                "image",
                "introductionVideo",
                "name",
                "prerequisite",
                "price",
                "thumbnailImage"
            ],
            "properties": {
                "abilityToAddComment": {
                    "type": "boolean"
                },
                "canHaveDiscount": {
                    "type": "boolean"
                },
                "categoryId": {
                    "type": "integer"
                },
                "commentAccessMode": {
                    "type": "string",
                    "enum": [
                        "all",
                        "students"
                    ]
                },
                "courseLevel": {
                    "type": "string",
                    "enum": [
                        "beginner",
                        "pre-intermediate",
                        "intermediate",
                        "advance"
                    ]
                },
                "description": {
                    "type": "string",
                    "minLength": 20
                },
                "image": {
                    "type": "string",
                    "minLength": 10
                },
                "introductionVideo": {
                    "type": "string",
                    "minLength": 20
                },
                "maxDiscountAmount": {
                    "type": "number",
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "prerequisite": {
                    "type": "string",
                    "minLength": 20
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnailImage": {
                    "type": "string",
                    "minLength": 10
                }
            }
        },
...
        "github_com_ladmakhi81_learnup_internals_teacher_dto_res.CreateCourseRes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_ladmakhi81_learnup_internals_teacher_dto_res.FetchCourseItemRes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "types.ApiError": {
            "type": "object",
            "properties": {
                "message": {},
                "statusCode": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "types.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "types.PaginationRes": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "row": {},
                "totalCount": {
                    "type": "integer"
                },
                "totalPage": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}